move actions under the same folder
