+++ src/actions/disableDraftMode.ts
@@ -1,12 +1,12 @@
-'use server'
+'use server';
 
-import { draftMode } from 'next/headers'
+import { draftMode } from 'next/headers';
 
 export async function disableDraftMode() {
-  'use server'
+  'use server';
   await Promise.allSettled([
     (await draftMode()).disable(),
     // Simulate a delay to show the loading state
     new Promise((resolve) => setTimeout(resolve, 1000)),
-  ])
+  ]);
 }
