+++ src/components/sections/Subscribe/SubscribeForm.tsx
@@ -1,11 +1,10 @@
 'use client';
 
-import subscribeAction from '@/actions/subscribeAction';
+import { subscribeAction } from '@/actions/subscribeAction';
 import { ActionResponse } from '@/actions/types';
 import { Badge } from '@/components/ui/Badge';
 import { Button } from '@/components/ui/Button';
 import type { SubscribeSectionFragmentType } from '@/lib/sanity/queries/fragments/fragment.types';
-
 import { useActionState } from 'react';
 
 const SubscribeForm = ({ section }: { section: SubscribeSectionFragmentType }) => {
